pmu/psm-firmware: Move from S to B which allows code to be patched
The system should never reference or build software that is above the 'S'
directory.  So S being the top of the source tree, and B (build) being the
location where we are running out build components will work correctly and
meet the Yocto Project requirements.

This also has a side effect of allowing patching of components above the
build directory, if necessary.

Signed-off-by: Mark Hatle <mark.hatle@kernel.crashing.org>